チャートスタイリングの修正
