+++ src/app/components/elements/authDetail/AuthDetail.module.scss
@@ -2,14 +2,14 @@
 
 .auth {
   display: flex;
-  justify-self: flex-start;
-  align-self: center;
+  justify-content: flex-start;
+  align-items: center;
   gap: 1rem;
   font-family: var(--font-en);
 
   &_icon_container {
-    width: 3rem;
-    height: 3rem;
+    width: 2rem;
+    height: 2rem;
     border-radius: 50%;
     position: relative;
     background-color: white;
