+++ src/app/components/elements/button/Button.module.scss
@@ -2,7 +2,7 @@
 
 .button {
   display: block;
-  width: 1.5rem;
+  width: 1.3rem;
   aspect-ratio: 1/1;
   position: relative;
