+++ src/app/components/elements/chart/Chart.module.scss
@@ -1,7 +1,7 @@
 @use '@/styles/_variables' as *;
 
 .skills_container {
-  width: 100%;
+  height: 100%;
   aspect-ratio: 1/1;
   position: relative;
 
@@ -22,16 +22,16 @@
     }
 
     @include mq(lg) {
-      width: 6rem;
+      width: 5.5rem;
     }
 
     &_name {
       display: inline-block;
       overflow: hidden;
       text-overflow: ellipsis;
-      width: 4rem;
+      width: 4.5rem;
       height: 1rem;
-      font-size: .8rem;
+      font-size: .7rem;
       font-weight: bold;
       color: white;
       white-space: nowrap;
@@ -39,7 +39,7 @@
 
       @include mq(md) {
         width: 5rem;
-        font-size: .9rem;
+        font-size: .8rem;
       }
     }
   }
