+++ src/app/components/layouts/nav/Nav.tsx
@@ -1,9 +1,12 @@
-import Link from "next/link";
 import Image from "next/image";
+import Link from "next/link";
+
 import IsHowToModalProvider from "@/components/utils/IsHowToModalProvider";
-import styles from "./nav.module.scss";
+import Logo from "@/components/elements/logo/Logo";
 import Logout from '../../elements/logout/Logout';
 import NavHowToButton from "./NavHowToButton";
+
+import styles from "./nav.module.scss";
 
 interface Props {
   props: string
@@ -21,6 +24,9 @@ export default function Nav({ props }: Props) {
 
   return (
     <nav className={styles.nav}>
+      <div className={styles.logo_container}>
+        <Logo size="lg" />
+      </div>
       <div className={styles.HowToButton_container}>
         <IsHowToModalProvider>
           <NavHowToButton />
