+++ src/app/components/layouts/nav/nav.module.scss
@@ -13,11 +13,12 @@
 
   @include mq(lg) {
     display: block;
-    width: 12.5rem;
+    width: 14rem;
     height: 100vh;
-    position: fixed;
+    padding: 1rem 3%;
+    position: absolute;
     top: 0;
-    left: 3%;
+    left: 0;
     background-color: transparent;
 
     &::before {
@@ -32,13 +33,22 @@
     }
   }
 
-  & .HowToButton_container{
+  & .logo_container {
+    display: none;
+
+    @include mq(lg) {
+      display: block;
+      margin-bottom: 2rem;
+    }
+  }
+
+  & .HowToButton_container {
     width: 10rem;
     display: none;
 
-    @include mq(lg){
+    @include mq(lg) {
       display: block;
-      padding: 5rem 0 0 .5rem;
+      margin-bottom: 2rem;
     }
   }
 
@@ -50,8 +60,7 @@
     place-items: center;
 
     @include mq(lg) {
-      height: calc(100vh - 10rem);
-      padding: 2rem 1rem 0 0;
+      height: calc(100vh - 15rem);
       grid-template-columns: 1fr;
       grid-template-rows: repeat(5, 1fr) 3fr;
     }
@@ -85,7 +94,6 @@
       @include mq(lg) {
         justify-content: start;
         gap: 1rem;
-        padding-left: 10%;
         position: relative;
 
         &::before {
@@ -138,7 +146,6 @@
   }
 
   &_logout {
-    // margin-top: 3rem;
-    padding-left: calc(10% + 2.5rem);
+    padding-left: 2.5rem;
   }
 }
