+++ src/app/features/chartView/styles/ChartDisp.module.scss
@@ -13,7 +13,7 @@
     font-size: 1rem;
 
     @include mq(lg) {
-      font-size: 1.3rem;
+      font-size: 1rem;
     }
   }
 
@@ -23,7 +23,7 @@
     font-size: 1rem;
 
     @include mq(lg) {
-      font-size: 1.3rem;
+      font-size: 1rem;
     }
   }
 }
