+++ src/app/features/chartView/styles/ChartIntro.module.scss
@@ -3,6 +3,7 @@
 .container {
   width: 100%;
   height: 100vh;
+  padding: 0 calc(20rem - 3%) 0 calc(14rem - 3%);
   display: flex;
   flex-direction: column;
   justify-content: center;
@@ -24,7 +25,7 @@
     border-radius: 10px;
     transition: all .1s $ease;
 
-    &:hover{
+    &:hover {
       background-color: lighten($green, 10%);
       transform: scale(1.1);
     }
