+++ src/app/features/chartView/styles/ChartSelect.module.scss
@@ -1,0 +1,2 @@
+@use '@/styles/_variables' as *;
+
