+++ src/app/styles/globals.scss
@@ -8,9 +8,10 @@
 
 html,
 body {
-  max-width: 100vw;
+  width: 100vw;
+  max-width: 1500px;
+  margin: 0 auto;
   height: 100%;
-  // padding-bottom: 3rem;
   overflow-x: hidden;
   @extend .font-md;
   font-family: var(--font-jp);
@@ -22,16 +23,6 @@ body {
 
 body {
   padding: 0 3%;
-  @include mq(lg) {
-  }
-}
-
-main{
-  @include mq(lg){
-    width: calc(100vw - 36rem);
-    margin: 0 auto;
-  }
-
 }
 
 a {
